+++ src/main/java/controller/ExceptionHandler.java
@@ -1,4 +1,4 @@
-package handler;
+package controller;
 
 import java.util.function.Supplier;
